+++ src/university_maps/core.clj
@@ -14,24 +14,31 @@
     (twitter.callbacks.protocols SyncSingleCallback))
   (:gen-class
     :name "UniversityData"
-    :methods [#^{:static true} [getData [String String] void]]))
+    :methods [#^{:static true} [getData [String String] void]
+              #^{:static true} [getData [String String int] void]]))
 
 (defn get-file-path
   "Given a file path, it gets all the files in the current directory
   and searches for the files."
   [file-name]
   (let [fs (file-seq (io/file (System/getProperty "user.dir")))
-        file-vec (vec fs)]
-    (.getPath (first (filter
-                       #(and (not (.isDirectory %))
-                             (= file-name (.getName %)))
-                       file-vec)))))
+        file-vec (vec fs)
+        file-loc (first (filter
+                          #(and (not (.isDirectory %))
+                                (= file-name (.getName %)))
+                          file-vec))]
+    (if (nil? file-loc)
+      nil
+      (.getPath file-loc))))
 
 ; Reads the data from UnivDataInfo.csv file. Data is split into 4 columsn for each
 ; university: name, latitude, longitude, hash-tag, and facebook-id. Each university is a hash-map
 ; inside an outer list
 (def university-data
-  (let [data-string (slurp (get-file-path "UnivDataInfo.csv"))
+  (let [file-loc (get-file-path "UnivDataInfo.csv")
+        data-string (if (nil? file-loc)
+                      ""
+                      (slurp file-loc))
         all-uni-vec (str/split data-string #"\r\n")
         all-individual-uni-vec (map #(str/split % #",") all-uni-vec)]
     (map #(zipmap [:uni-name :lat :long :uni-hash-tag :uni-facebook-id] %) all-individual-uni-vec)))
@@ -40,7 +47,10 @@
 ; word: Word and the corresponding sentiment value. Each word is a hashtag inside an
 ; outer list
 (def word-sentiment-data
-  (let [line (slurp (get-file-path "WordSentiment.csv"))
+  (let [file-loc (get-file-path "WordSentiment.csv")
+        line (if (nil? file-loc)
+               ","
+               (slurp file-loc))
         all-word-vector (str/split line #"\n")]
     (apply hash-map (flatten (map #(str/split % #",") all-word-vector)))))
 
@@ -64,13 +74,13 @@
 (defn search-university-by-hashtag
   "Searches twitter given a hashtag and returns the JSON response.
   It gets ten of the most recent english responses back"
-  [uni-hashtag]
+  [uni-hashtag num-tweets]
   (search-tweets :oauth-creds my-creds
                  :callbacks (SyncSingleCallback. response-return-body
                                                  response-throw-error
                                                  exception-rethrow)
                  :params {:q                uni-hashtag
-                          :count            10
+                          :count            num-tweets
                           :lang             "en"
                           :result-type      "recent"
                           :include-entities false}))
@@ -78,8 +88,8 @@
 (defn get-university-tweets
   "Based on the hash-tag, it gets the responses puts the each text from the tweet
   in a list of strings"
-  [uni-hashtag]
-  (->> (get-in (search-university-by-hashtag uni-hashtag) [:statuses])
+  [uni-hashtag num-tweets]
+  (->> (get-in (search-university-by-hashtag uni-hashtag num-tweets) [:statuses])
        (map #(get % :text))))
 
 ; Creates a facebook authentication of an app access token
@@ -89,10 +99,12 @@
   "Based on the page id, it gets the list of 10 posts on the page's feed.
   Facebook Graph API returns a list of JSON Responses and it only keeps a
   list of the messages on the page's feed."
-  [page-id]
+  [page-id num-posts]
   (map
     #(get % :message)
-    (with-facebook-auth facebook-auth (client/get [page-id :feed] {:query-params {:limit 10} :extract :data}))))
+    (with-facebook-auth
+      facebook-auth
+      (client/get [page-id :feed] {:query-params {:limit num-posts} :extract :data}))))
 
 (defn write-to-file
   "Takes in a file location a list of the lines to be put into the file.
@@ -108,7 +120,7 @@
   [uni text]
   (let [line (if (nil? text)
                ""
-               (str/replace text #"[.,?!\r\n\t]" ""))]
+               (str/replace text #"[.,?!'\"\r\n\t]" ""))]
     (str (get uni :uni-name) "," (get uni :lat) "," (get uni :long) "," (get uni :uni-hash-tag) "," (get uni :uni-facebook-id)
          "," line "," (sum-sentiment line) "\n")))
 
@@ -124,40 +136,49 @@
   "For each university, it gets the tweets and facebook page feed,
   calculates the sentiment for each post, and puts each line back
   into the respective csv files"
-  [twitter-location facebook-location]
-  (do
-    (create-file twitter-location)
-    (create-file facebook-location)
-    (println "Start")
-    (doseq [uni university-data]
-      (println (str "Getting " (get uni :uni-name) "'s data"))
-      (let [twitter-data (->> (get uni :uni-hash-tag)
-                              (get-university-tweets)
-                              (map #(create-line uni %))
-                              (future))
-            facebook-data (->> (get uni :uni-facebook-id)
-                               (get-facebook-page-feed)
-                               (map #(create-line uni %))
-                               (future))]
-        (do
-          (write-to-file twitter-location @twitter-data)
-          (write-to-file facebook-location @facebook-data))))
-    (println "Done")))
+  ([twitter-location facebook-location]
+    (get-data twitter-location facebook-location 10))
+  ([twitter-location facebook-location num-results]
+   (if (or (< num-results 1) (= (count university-data) 1))
+     (println "Error: Cannot get the data. Check to make sure both
+    UnivDataInfo.csv and WordSentiment.csv are in the directory. Also
+    check to make sure you are asking for one or more tweets/posts per university")
+     (do
+       (create-file twitter-location)
+       (create-file facebook-location)
+       (println "Start")
+       (doseq [uni university-data]
+         (println (str "Getting " (get uni :uni-name) "'s data"))
+         (let [twitter-data (->> (get uni :uni-hash-tag)
+                                 (get-university-tweets num-results)
+                                 (map #(create-line uni %))
+                                 (future))
+               facebook-data (->> (get uni :uni-facebook-id)
+                                  (get-facebook-page-feed num-results)
+                                  (map #(create-line uni %))
+                                  (future))]
+           (write-to-file twitter-location @twitter-data)
+           (write-to-file facebook-location @facebook-data)))
+       (println "Done")))))
 
 (defn -getData
   "Simple wrapper function for Java class that can call Clojure function (get-data)
   Takes in two parameters which determine the locations to store the Twitter and Facebook
   data respectively"
-  [twitter-location facebook-location]
-  (if (or (nil? twitter-location) (nil? facebook-location))
-    (get-data "output/TwitterOutput.csv" "output/FacebookOutput.csv")
-    (get-data twitter-location facebook-location)))
+  ([twitter-location facebook-location]
+    (-getData twitter-location facebook-location 10))
+  ([twitter-location facebook-location num-results]
+   (if (or (nil? twitter-location) (nil? facebook-location) (< num-results 1))
+     (get-data "output/TwitterOutput.csv" "output/FacebookOutput.csv" 10)
+     (get-data twitter-location facebook-location num-results))))
 
 (defn -main
   "Main function that will be called when the Jar file is called by itself.
-  If two locations are not provided, it will output to a default directory (called output)"
+  If two locations are not provided, it will output to a default directory (called output)
+  with 10 of the most recent tweets/posts from each university. Otherwise, it
+  it outputs to the desired place with the correct number of tweets/posts for each university"
   [& args]
-  (if (or (< (count args) 2)
-          (or (nil? (first args)) (nil? (second args))))
+  (if (or (< (count args) 3)
+          (or (nil? (first args)) (nil? (second args)) (< (nth args 3) 1)))
     (get-data "output/TwitterOutput.csv" "output/FacebookOutput.csv")
-    (get-data (first args) (second args))))
+    (get-data (first args) (second args) (nth args 3))))